grIceMt_801F9ACC: s16->s32 id resolves register-coloring wall (98.84->98.95%)
The directed-search pilot's order-distance metric found that widening the local
'id' (a segment index, assigned from seg[] and only compared !=-1 / passed to
Ground_801C2BA4) from s16 to s32 shifts the colorgraph simplify order so ig40/gp
colors 3rd (r29) and ig33/y colors 5th (r27) — the target coloring. checkdiff
confirms: the register swaps are gone from the diff (+0.11%). 'int/s32 id' is the
natural type for an index; the s16 was a decomp mis-guess. 72K permuter iters +
manual levers missed this. Remaining residual: named-float .sdata2 relocs + one
fmr/lha schedule swap.

+++ src/melee/gr/gricemt.c
@@ -1474,7 +1474,7 @@ int grIceMt_801F9ACC(Ground_GObj* gobj, float y, GrIceMtSegmentLookup ev,
     f32 f;
     f32 f2;
     HSD_JObj** new_var;
-    s16 id;
+    s32 id;
     PAD_STACK(16);
 
     if (seg[0] == -1 || seg[1] == -1) {
